requirements.txt
Flask == 0.12.2
azure-cosmos==3.1.1
pytest==4.6.3
